+++ src/main/framework/routes/factory-routes.ts
@@ -1,3 +1,4 @@
+import { AuthRouter } from './auth-router';
 import { PurchaseListRoutes } from './create-purchase-list-router';
 import { SignupRouter } from './signup-router';
 
@@ -5,10 +6,12 @@ export class FactoryRoutes {
   static createRoutes() {
     const signupRoutes = new SignupRouter();
     const purchaseListRoutes = new PurchaseListRoutes();
+    const authRouter = new AuthRouter();
 
     return {
       signupRoutes: signupRoutes.getRouter(),
       purchaseListRoutes: purchaseListRoutes.getRouter(),
+      authRoutes: authRouter.getRouter(),
     };
   }
 }
